+++ examples/o3r/configuration/configuration.cpp
@@ -8,7 +8,7 @@
 #include <memory>
 #include <fstream>
 #include <ifm3d/device/o3r.h>
-using json = nlohmann::json;
+using json = ifm3d::json;
 
 
 int main(){
@@ -23,7 +23,7 @@ int main(){
     std::cout << std::setw(4) << conf << std::endl;
     
     // Configure the device from a json string
-    cam->Set(json::parse(R"({"device":{"info":{"name": "my_new_o3r"}}})")); 
+    cam->Set(ifm3d::json::parse(R"({"device":{"info":{"name": "my_new_o3r"}}})")); 
     // Make the configuration persistent
     cam->SaveInit();