+++ examples/o3r/getting_data/getting_data.cpp
@@ -8,6 +8,7 @@
 #include <ifm3d/fg.h>
 
 using namespace std::chrono_literals;
+using namespace ifm3d::literals;
 
 int
 main()